+++ src/main/java/Junny/Deadline.java
@@ -1,3 +1,5 @@
+package Junny;
+
 import java.time.LocalDate;
 import java.time.format.DateTimeFormatter;
