+++ src/main/java/Junny/Event.java
@@ -1,4 +1,5 @@
-import javax.swing.text.DateFormatter;
+package Junny;
+
 import java.time.LocalDate;
 import java.time.format.DateTimeFormatter;
