+++ src/main/java/Junny/ListCommand.java
@@ -1,6 +1,8 @@
+package Junny;
+
 import java.util.ArrayList;
 
-public class ListCommand extends Command{
+public class ListCommand extends Command {
     @Override
     public void run(ArrayList<Task> tasks, Ui ui, Storage storage){
         ui.printAllTasks(tasks, tasks.size());
