+++ src/main/java/Junny/Parser.java
@@ -1,3 +1,5 @@
+package Junny;
+
 import java.time.LocalDate;
 
 public class Parser {
