+++ src/main/java/Junny/Storage.java
@@ -1,5 +1,5 @@
-import java.io.File;
-import java.io.FileWriter;
+package Junny;
+
 import java.io.IOException;
 import java.nio.file.Files;
 import java.nio.file.Path;
