+++ src/main/java/Junny/Task.java
@@ -1,3 +1,6 @@
+package Junny;
+
+
 public abstract class Task {
     protected String description;
     protected boolean isDone;
