+++ src/main/java/Junny/TaskTypes.java
@@ -1,3 +1,5 @@
+package Junny;
+
 public enum TaskTypes {
     TODO,
     DEADLINE,
