+++ src/main/java/Junny/Todo.java
@@ -1,4 +1,6 @@
-public class Todo extends Task{
+package Junny;
+
+public class Todo extends Task {
     public Todo(String description) {
         super(description, TaskTypes.EVENT);
     }
