+++ src/main/java/Junny/possible commands
