+++ text-ui-test/output.TXT
@@ -1,5 +1,5 @@
 ____________________________________________________________
- Hello! I'm Junny
+ Hello! I'm Junny.Junny
  What can I do for you?
 ____________________________________________________________
 ____________________________________________________________
